requirements.txt
aiogram
firebase-admin
python-dotenv
requests